--- .prettierrc
@@ -1,11 +1,0 @@
-{
-  "allowParens": "always",
-  "bracketSpacing": true,
-  "printWidth": 150,
-  "proseWrap": "preserve",
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "es5",
-  "useTabs": false,
-  "tabWidth": 2
-}
