--- src/css-variabale-generator/.eslintrc.json
@@ -1,19 +1,0 @@
-{
-  "root": true,
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaVersion": 6,
-    "sourceType": "module"
-  },
-  "plugins": ["@typescript-eslint"],
-  "rules": {
-    "@typescript-eslint/naming-convention": "warn",
-    "@typescript-eslint/semi": "warn",
-    "curly": "warn",
-    "eqeqeq": "warn",
-    "no-throw-literal": "warn",
-    "semi": "off",
-    "object-property-newline": ["error", { "allowAllPropertiesOnSameLine": true }]
-  },
-  "ignorePatterns": ["out", "dist", "**/*.d.ts"]
-}
