--- src/css-variabale-generator/.gitignore
@@ -1,37 +1,0 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
-
-# config files
-
-.editorconfig
-.eslintrc
-.jshintrc
-.lesshintrc
-
-# dependencies
-/node_modules/*
-/bower_components/*
-
-# testing
-/coverage/*
-/dist/*
-/.tmp/*
-# production
-/build/*
-
-# misc
-.DS_Store
-**/.DS_Store
-.env.local
-Thumbs.db
-desktop.ini
-
-# IDE
-*.sublime-*
-*.code-workspace
-/.idea
-*~
-
-**/log/
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
