--- src/css-variabale-generator/.vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-	// See http://go.microsoft.com/fwlink/?LinkId=827846
-	// for the documentation about the extensions.json format
-	"recommendations": [
-		"dbaeumer.vscode-eslint"
-	]
-}
