--- src/css-variabale-generator/.vscode/settings.json
@@ -1,14 +1,0 @@
-// Place your settings in this file to overwrite default and user settings.
-{
-  "files.exclude": {
-    "out": false // set this to true to hide the "out" folder with the compiled JS files
-  },
-  "search.exclude": {
-    "out": true // set this to false to include "out" folder in search results
-  },
-  // Turn off tsc task auto detection since we have the necessary tasks as npm scripts
-  "typescript.tsc.autoDetect": "off",
-  "eslint.alwaysShowStatus": true,
-  "eslint.debug": true,
-  "eslint.format.enable": true
-}
