--- src/css-variabale-generator/src/models/base.ts
@@ -1,5 +1,0 @@
-export type VariableList = {
-	[key: string]: string
-};
-
-export type RegExpMatchArrayWithIndices = RegExpMatchArray & { indices: Array<[number, number]> & { groups: { [key: string]: [number, number] } } };
