--- src/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "src",
-  "version": "1.0.0",
-  "description": "vs-code extension",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/xkeshav/parixan.git"
-  },
-  "keywords": [
-    "vs",
-    "code"
-  ],
-  "author": "Keshav Mohta",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/xkeshav/parixan/issues"
-  },
-  "homepage": "https://github.com/xkeshav/parixan#readme"
-}
