--- src/test/styles/base.css
@@ -1,32 +1,0 @@
-.nav {
-  display: grid;
-  width: 40vw;
-  height: 20vh;
-  margin: 2em 4rem;
-  background-color: aliceblue;
-  transition: background-color 0.1s ease-in-out;
-  place-items: center;
-  will-change: contents;
-}
-
-.nav:hover {
-  background-color: blanchedalmond;
-}
-
-.section--content {
-  border: 1px dashed gainsboro;
-  margin: 0.5rem;
-  padding: 1rem;
-  font-size: medium;
-  height: 1rem;
-  mix-blend-mode: difference;
-  background: linear-gradient(to left, #706363 20%, #fcfffe);
-}
-
-.red {
-  background-color: red;
-}
-
-.rdb {
-  border: 1px solid red;
-}
