--- src/test/styles/main.css
@@ -1,30 +1,0 @@
-@import url('./responsive.css');
-
-.grid--main {
-  display: grid;
-  grid-auto-columns: 1fr;
-  grid-auto-flow: dense;
-  grid-gap: 0.5rem 0;
-  grid-template-columns: repeat(12, [content-start] 1fr [content-end]);
-  grid-template-rows: [body-start] 1fr [body-end foot-start] minmax(var(--footer-height), auto) [foot-end];
-  min-height: calc(100vh var(--header-height));
-}
-
-.grid--body {
-  grid-column: var(--column-spread);
-  grid-row: body;
-  margin: var(--content-margin, 0);
-}
-
-.grid--foot {
-  border: 1px solid var(--border-color, grey);
-  grid-row: foot;
-  grid-column: var(--column-spread);
-  margin: var(--content-margin, 0);
-}
-
-/* when no footer */
-
-.grid--body--only {
-  grid-row: body / span 2;
-}
