--- src/test/styles/responsive.css
@@ -1,32 +1,0 @@
-:root {
-  --footer-height: 7rem;
-  --header-height: 7.25rem;
-  --content-margin: 0;
-  --column-spread: content-start / span 12; /* default spread till full 12 columns*/
-  --input-spread: input-start 1 / span 12;
-
-  /** css variable not works in @media syntax */
-  --device-small: 641px; /* iPad, tablet */
-  --device-medium: 961px; /* iPad, tablet in landscape*/
-  --device-large: 1281px; /* iPad pro in landscape*/
-}
-
-@media all and (min-width: 641px) {
-  :root {
-    --content-margin: 0 0.25rem;
-    --column-spread: content-start 3 / span 8; /* spread till 8 column */
-    --input-spread: input-start 1 / span 3;
-  }
-}
-
-@media all and (min-width: 961px) {
-  :root {
-    --input-spread: input-start 1 / span 9;
-  }
-}
-
-@media all and (min-width: 1281px) {
-  :root {
-    --input-spread: input-start 1 / span 6;
-  }
-}
